Refactor DRY
