--- utils/capitalize.js
@@ -1,5 +1,0 @@
-function capitalize (string) {
-  return string[0].toUpperCase() + string.slice(1).toLowerCase();
-}
-
-module.exports = capitalize;
